+++ integrationtest/src/test/java/dk/medcom/healthcheck/integrationtest/AbstractIntegrationTest.java
@@ -1,0 +1,54 @@
+package dk.medcom.healthcheck.integrationtest;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.testcontainers.containers.GenericContainer;
+
+import java.io.IOException;
+import java.net.URISyntaxException;
+
+public abstract class AbstractIntegrationTest {
+    private static final Logger logger = LoggerFactory.getLogger(AbstractIntegrationTest.class);
+
+    private static GenericContainer helloService;
+    private static String apiBasePath;
+
+    static {
+        Runtime.getRuntime().addShutdownHook(new Thread()
+        {
+            public void run()
+            {
+                if(helloService != null) {
+                    logger.info("Stopping hello service container: " + helloService.getContainerId());
+                    helloService.getDockerClient().stopContainerCmd(helloService.getContainerId()).exec();
+                }
+            }
+        });
+
+        try {
+            setup();
+        } catch (IOException | URISyntaxException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private static void setup() throws IOException, URISyntaxException {
+        var runInDocker = Boolean.getBoolean("runInDocker");
+        logger.info("Running integration test in docker container: " + runInDocker);
+
+        ServiceStarter serviceStarter;
+        serviceStarter = new ServiceStarter();
+        if(runInDocker) {
+            helloService = serviceStarter.startServicesInDocker();
+            apiBasePath = "http://" + helloService.getContainerIpAddress() + ":" + helloService.getMappedPort(8080);
+        }
+        else {
+            serviceStarter.startServices();
+            apiBasePath = "http://localhost:8080";
+        }
+    }
+
+    String getApiBasePath() {
+        return apiBasePath;
+    }
+}
