+++ integrationtest/src/test/java/dk/medcom/healthcheck/integrationtest/TestApplication.java
@@ -1,0 +1,11 @@
+package dk.medcom.healthcheck.integrationtest;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+
+@SpringBootApplication
+public class TestApplication extends SpringBootServletInitializer {
+    public static void main(String[] args) {
+        new ServiceStarter().startServices();
+    }
+}
