--- service/src/main/java/dk/kvalitetsit/hello/configuration/DatabaseConfiguration.java
@@ -1,30 +1,0 @@
-package dk.kvalitetsit.hello.configuration;
-
-import dk.kvalitetsit.hello.dao.HelloDao;
-import dk.kvalitetsit.hello.dao.HelloDaoImpl;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.jdbc.datasource.DriverManagerDataSource;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
-
-import javax.sql.DataSource;
-
-@Configuration
-@EnableTransactionManagement
-public class DatabaseConfiguration {
-    @Bean
-    public HelloDao helloDao(DataSource dataSource) {
-        return new HelloDaoImpl(dataSource);
-    }
-
-    @Bean
-    public DataSource dataSource(@Value("${jdbc.url}") String jdbcUrl, @Value("${jdbc.user}") String jdbcUser, @Value("${jdbc.pass}") String jdbcPass) {
-        DriverManagerDataSource dataSource = new DriverManagerDataSource();
-        dataSource.setUrl(jdbcUrl);
-        dataSource.setUsername(jdbcUser);
-        dataSource.setPassword(jdbcPass);
-
-        return dataSource;
-    }
-}
