--- service/src/main/java/dk/kvalitetsit/hello/configuration/HelloConfiguration.java
@@ -1,14 +1,0 @@
-package dk.kvalitetsit.hello.configuration;
-
-import dk.kvalitetsit.hello.service.HelloService;
-import dk.kvalitetsit.hello.service.HelloServiceImpl;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class HelloConfiguration{
-    @Bean
-    public HelloService helloService() {
-        return new HelloServiceImpl();
-    }
-}
