--- service/src/main/java/dk/kvalitetsit/hello/controller/exception/ResourceNotFoundException.java
@@ -1,11 +1,0 @@
-package dk.kvalitetsit.hello.controller.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.NOT_FOUND)
-public class ResourceNotFoundException extends RuntimeException {
-    public ResourceNotFoundException(String message) {
-        super(message);
-    }
-}
