--- service/src/main/java/dk/kvalitetsit/hello/service/HelloService.java
@@ -1,8 +1,0 @@
-package dk.kvalitetsit.hello.service;
-
-import dk.kvalitetsit.hello.service.model.HelloServiceOutput;
-import dk.kvalitetsit.hello.service.model.HelloServiceInput;
-
-public interface HelloService {
-    HelloServiceOutput helloServiceBusinessLogic(HelloServiceInput input);
-}
