--- service/src/main/java/dk/kvalitetsit/hello/service/model/HelloServiceOutput.java
@@ -1,6 +1,0 @@
-package dk.kvalitetsit.hello.service.model;
-
-import java.time.ZonedDateTime;
-
-public record HelloServiceOutput(String name, ZonedDateTime now) {
-}
