+++ service/src/main/java/dk/medcom/healthcheck/controller/exception/BadRequestException.java
@@ -1,0 +1,8 @@
+package dk.medcom.healthcheck.controller.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value = HttpStatus.BAD_REQUEST)
+public class BadRequestException extends RuntimeException {
+}
