+++ service/src/main/java/dk/medcom/healthcheck/controller/exception/ResourceNotFoundException.java
@@ -1,0 +1,11 @@
+package dk.medcom.healthcheck.controller.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value = HttpStatus.NOT_FOUND)
+public class ResourceNotFoundException extends RuntimeException {
+    public ResourceNotFoundException(String message) {
+        super(message);
+    }
+}
