+++ service/src/main/java/dk/medcom/healthcheck/dao/HelloDao.java
@@ -1,0 +1,11 @@
+package dk.medcom.healthcheck.dao;
+
+import dk.medcom.healthcheck.dao.entity.HelloEntity;
+
+import java.util.List;
+
+public interface HelloDao {
+    void insert(HelloEntity helloEntity);
+
+    List<HelloEntity> findAll();
+}
