+++ service/src/main/java/dk/medcom/healthcheck/service/HelloService.java
@@ -1,0 +1,8 @@
+package dk.medcom.healthcheck.service;
+
+import dk.medcom.healthcheck.service.model.HelloServiceOutput;
+import dk.medcom.healthcheck.service.model.HelloServiceInput;
+
+public interface HelloService {
+    HelloServiceOutput helloServiceBusinessLogic(HelloServiceInput input);
+}
