+++ service/src/main/java/dk/medcom/healthcheck/service/model/HelloServiceInput.java
@@ -1,0 +1,4 @@
+package dk.medcom.healthcheck.service.model;
+
+public record HelloServiceInput(String name) {
+}
