+++ service/src/main/java/dk/medcom/healthcheck/service/model/HelloServiceOutput.java
@@ -1,0 +1,6 @@
+package dk.medcom.healthcheck.service.model;
+
+import java.time.ZonedDateTime;
+
+public record HelloServiceOutput(String name, ZonedDateTime now) {
+}
