+++ service/src/test/java/dk/medcom/healthcheck/service/HelloServiceImplTest.java
@@ -1,0 +1,29 @@
+package dk.medcom.healthcheck.service;
+
+import dk.medcom.healthcheck.service.model.HelloServiceInput;
+import org.junit.Before;
+import org.junit.Test;
+
+import java.util.UUID;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+
+public class HelloServiceImplTest {
+    private HelloService helloService;
+
+    @Before
+    public void setup() {
+        helloService = new HelloServiceImpl();
+    }
+
+    @Test
+    public void testValidInput() {
+        var input = new HelloServiceInput(UUID.randomUUID().toString());
+
+        var result = helloService.helloServiceBusinessLogic(input);
+        assertNotNull(result);
+        assertNotNull(result.now());
+        assertEquals(input.name(), result.name());
+    }
+}
