--- web/src/main/java/dk/kvalitetsit/hello/VideoLinkHandlerApplication.java
@@ -1,11 +1,0 @@
-package dk.kvalitetsit.hello;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class VideoLinkHandlerApplication {
-	public static void main(String[] args) {
-		SpringApplication.run(VideoLinkHandlerApplication.class, args);
-	}
-}
