+++ web/src/main/java/dk/medcom/healthcheck/VideoLinkHandlerApplication.java
@@ -1,0 +1,11 @@
+package dk.medcom.healthcheck;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class VideoLinkHandlerApplication {
+	public static void main(String[] args) {
+		SpringApplication.run(VideoLinkHandlerApplication.class, args);
+	}
+}
