+++ web/src/test/java/dk/medcom/healthcheck/VideoLinkHandlerApplicationTests.java
@@ -1,0 +1,14 @@
+package dk.medcom.healthcheck;
+
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+@Disabled
+class VideoLinkHandlerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+}
